requirements.txt
Django==5.0
psycopg[binary]
python-dateutil